Do not ignore SubmitRecord status from Prolog rules
When translating a SubmitRecord into a SubmitRequirement in Ie0fbf0e70,
the status field was forgotten in the translation and effectively ignored.
A custom Prolog rule returning a status "OK" with extra labels was
therefore considered as not satisifed requirement whilst it should have
been passed them.

When the overall SubmitRecord.status is OK, skip all
the other labels that are not compatible with the
SubmitRecord.status field.

Restore the behaviour observed in Gerrit with custom
Prolog rules in v3.5 or earlier versions.

Bug: Issue 343218480
Release-Notes: Fix translation of custom Prolog rules SubmitRecord to SubmitRequirement(s)
Change-Id: Ib0ac6c7638dece3d14261a3ac45c43825e7ecec8