Increase maximum weight for the project list cache to fix data race
The project list cache uses Caffeine. For Caffeine caches the configured
maximum weight may be temporarily exceeded while evicting. If this
happens Caffeine increases the cache size in a non-thread-safe way which
can cause a data race. To prevent this from happening we must set the
maximum weight high enough so that resizing during eviction is not
necessary. From internal tests that detect data races we see that the
cache size may be increased to 4, and setting 4 as the maximum weight
fixes the data race.

This problem has been reported to the Caffeine team:
https://github.com/ben-manes/caffeine/issues/1698

Release-Notes: skip
Bug: Google b/306907433
Change-Id: I7581f3b73ebf9027011ab8a9033447d98e75b661
Signed-off-by: Edwin Kempin <ekempin@google.com>