Bazel: Don't build protobuf from the source
Gerrit is pure java project. Given that it depends on Google protobuf,
and given that Bazel was using @com_google_protobuf toolchain that was
built from source, this project ended up bulding the Google protobuf
from source.

That changed in Bazel 7.x release, with the new and shiny option:
--incompatible_enable_proto_toolchain_resolution, that allow to register
prebuilt protoc toolchains.

In addition rules_proto have added support for prebuilt toolchains in
context of this tracking issue: [1].

In this change we use toolchains_protoc project for predefined protobuf
toolchains.

Now, that we stop building protobuf from source, we can remove C++
options in .bazelrc as well.

[1] https://github.com/bazelbuild/rules_proto/issues/179
[2] https://github.com/aspect-build/toolchains_protoc

Release-Notes: Use prebuilt
Change-Id: I27975879819c4b632682990474ce88737f722d9a