POC of StaticItem binding for plugins
In some cases plugins would like to expose their API to other plugins,
for example to give access to internal state or to programatically
modify its configuraiton.

This can be potentially done currently with the usage of DynamicItem,
but that doesn't guarantee that the implemenation will be bound from a
given plugin. Leaving potential for missuse of the binding from other
plugin.

The StaticItem is an attempt to solve that limitation. With
`StaticItem.itemOf()` not only accepting the `Binder` and item class,
but also name of the plugin that must provide the implementaiton.

To use the new StaticItem, bind the item type and implementaion plugin
name in the `ApiModule`:

```java
StaticItem.itemOf(binder(), MyPluginApi.class, "my-plugin");
```

Then in the main plugin Guice module bind the implementaion:

```java
StaticItem.bind(binder(), MyPluginApi.class).to(MyPluginApiImpl.class);
```

Finally, to consume the implementaion, in any plugin inject:

```java
StaticItem<MyPluginApi> myPluginApiItem
```

Then in the plugin code use `StaticItem.get()` method to obtain the
implementation.

Change-Id: Idec3985b2dd2dc753a88db7366cabec9a469f880