[DO NOT MERGE] Introduce rememberMe in login for become any account
When using the DEVELOPMENT_BECOME_ANY_ACCOUNT auth type, it may
be useful to enable the 'rememberMe' option on the cookie by
adding the Java System Property 'ghs.gerrit.auth.remember-me' to true.

The persistence is also associated with the XSRF_TOKEN with the same
validity of the GerritAccount cookie.

Change-Id: I85b4f9de9736cabc1dce9a58efa0aa29a7e4a39d