Fix updating an existing edit via push
It should be possible to update an existing edit via push [1], but
attempting to so (checkout the edit, amend the commit and then push the
edit) fails because updating the edit ref with non-fastword is rejected.

By setting the type to UPDATE_NONFASTFORWARD when creating the
ReceiveCommand for the edit ref update we allow non-fastforwards (see
NoteDbUpdateExecutor#allowNonFastForwards).

[1] https://gerrit-review.googlesource.com/Documentation/user-upload.html#:~:text=When%20a%20change%20edit%20already%20exists%20for%20a%20change%20then%20pushing%20with%20%25edit%20replaces%20the%20existing%20change%20edit.%20This%20option%20is%20useful%20to%20rebase%20a%20change%20edit%20on%20the%20newest%20patch%20set%20when%20the%20rebase%20of%20the%20change%20edit%20in%20the%20web%20UI%20fails%20due%20to%20conflicts.

Release-Notes: Fixed updating an existing edit via push
Bugs: Google b/332704794
Change-Id: Ia609a3c4ef8f949981644f2b36775e962359de48
Signed-off-by: Edwin Kempin <ekempin@google.com>