Only show matching hashtags as suggests
The current behavior is to show all hashtags from changes that has one
hashtag that matches input. For a server that has wide use of hashtags
this is less helpful since a majority of the suggested hashtags aren't
matching the input.
Filter the array of existing hashtags for those matching the input.

Solves: bug 344608194
Change-Id: I7d75184594e925890860c84c156aee86eb782a08