Update PluginApi.hook() exposed API to PluginElement
The TypeScript API published on npm exposes `PluginApi.hook` with a
constraint on HTMLElement.  The implementation in gr-public-js-api.ts
uses a `PluginElement` which exposes a few more properties such as
`change` or `revision`.

Thus, when using the hook and passing the following code through tsc:

  plugin.hook('commit-container').onAttached( element => {
    element.change;
  } );

TypeScript complains with:

  error TS2551: Property 'change' does not exist on type 'HTMLElement'.
  Did you mean 'onchange'?

Update the TypeScript API to match the implementation, changing
PluginApi.hook() to use the child class PluginElement instead of
HTMLElement.

Release-Notes: TypeScript API: fix PluginApi.hook() exposed type from HTMLElement to PluginElement
Change-Id: Ia96c71bbfc11c30e76c121c24cc6d0b9eec819c7