Compute diff URLs in the view model
This fixes a bug where `checksPatchset` would be reset when navigating
to a diff view.

Generally, when we are constsructing a diff URL we are typically
interested into computing a URL relative to the current state. We do
not want to look up and set all current view properties in every
code location that creates a diff URL. So we move the construction
of diff URLs into the view model itself and just provide
the delta as an override.

Release-Notes: skip
Google-Bug-Id: b/345851627
Change-Id: Ib13529a50c8a952e41eb605e01f654b31dace5c6