Compute diff URLs in the view model
This rolls forward change 429437. Compare against patchset 1 to see
what has changed compared to the version of the change that had to be
rolled back. The reason for rolling back was that we actually have to
provide the `patchNum` to the view model. For diff views the specific
patchNum has to be provided, even if your choice is the "latest" patch
num. That is different from the change overview page, which allows the
patchNum in the URL to be undefined.

Original commit message:

This fixes a bug where `checksPatchset` would be reset when navigating
to a diff view.

Generally, when we are constsructing a diff URL we are typically
interested into computing a URL relative to the current state. We do
not want to look up and set all current view properties in every
code location that creates a diff URL. So we move the construction
of diff URLs into the view model itself and just provide
the delta as an override.

Release-Notes: skip
Google-Bug-Id: b/345851627
Change-Id: If874a91d05800a06efe800e656567134915b3013