Remove etags for change resources
Since change etags are expensive to compute and their value was
questionable change Ib1ca47318 added an experiment for disabling change
etags (see commit message of that change for details). Google run this
experiment in the prod environment for 1 week and we carefully observed
the latency of change detail requests. After the 1 week we can confirm
that etags do not have a significant impact on performance. Actually the
performance of change detail calls from PolyGerrit slightly improved
after removing change etags. Over all change detail requests there was a
very slight performance regression, but the regression is too small to
tell for sure whether it was caused by the removal of change etags (or
something else in the system has caused it).

Since there is no real performance benefit of having change etag we
decided to remove them. This has the following benefits:
* Makes requests from PolyGerrit slightly faster.
* Reduces code complexity by removing the etag computation code.
* Fixes some edge cases that are currently not covered in the etag
  computation (e.g. the etag was not updated when changes to group
  memberships of reviewers happened so that their permitted labels
  didn't get updated).

In this change we can't drop all of the code for the change etag
computation yet, because it's still used by the revision etag
computation, but in a follow-up change we likely want to drop that etag
computation too.

Release-Notes: Removed etags for change resources
Bug: Google b/336400432
Change-Id: Ia92fbad2ad187d4fcb6fe237b364d33a5675c5cd
Signed-off-by: Edwin Kempin <ekempin@google.com>