Remove etags for revision resources
Change Ia92fbad2a removes etags for change resources because the change
etag computation is too expensive to bring any latency benefits. The
same is true for the revision etag, i.e. it includes the expensive
change etag computation.

An etag is only returned if the resource implements the HasETag
interface and if the REST endpoint returns a response with cache
control, see RestApiServlet#configureCaching which calls
RestApiServlet#addResourceStateHeaders that sets the etag in the
response. For the revision REST endpoints only the Get Review REST
endpoint returns a response with cache control, so only for this REST
endpoint a revision etag is returned.

Etags are mostly relevant for UI's that do auto-refreshing via polling,
but PolyGerrit doesn't use the Get Review REST endpoint, and hence
doesn't make use of this etag. For other callers (mostly service users)
it's unlikely that they do polling with etags, especially since the etag
support for this REST endpoint is not even documented [1]. Hence having
revision etags likely doesn't have any benefits, but calculating them
can add up to 1.2s to Get Review requests, just for including the etag
into the response.

In this change we still can't drop all of the code for the change etag
and revision etag computation yet, because it's still used by the etag
computation for ListFiles, but in a follow-up change we likely want to
drop that etag computation too.

[1] https://gerrit-review.googlesource.com/Documentation/rest-api-changes.html#get-review

Release-Notes: Removed etags for revision resources
Bug: Google b/336400432
Change-Id: Ic54db5dc96614a1dafef5c777828156fb1be2209
Signed-off-by: Edwin Kempin <ekempin@google.com>