Add ChangeKind for trivial rebase with commit message update.
ChangeKind is used when defining rules, that control whether or not the
vote is copied, after the new patchset have been uploaded. It's common
configuration for both TRIVIAL_REBASE and NO_CODE_CHANGE (commit message
update) to not invalidate votes. Adding this extra value allows to
changes that do both actions to also not invalidate votes. While in
principle such change can be split in two, sometimes the edit to commit
message is the result of the tooling and therefore not easily avoidable.

At the same time, we don't want to simply change the definition of the
TRIVIAL_REBASE, since some repos might prefer commit changes to cause a
re-review.

Additionally we add the unit tests to the ChangeKindCacheImpl for both
new and old usecases, since they were missing before.

Also update the ChangeKind returned, when both `prior` and `next` are
the same ObjectId. The only reason why it's NO_CODE_CHANGE currently is
because, at the time it was added (Ie04d0282e) NO_CHANGE didn't yet
exist, and it hasn't been adjusted since.

Google-Bug-Id: b/324626345
Release-Notes: Add TRIVIAL_REBASE_WITH_MESSAGE_UPDATE ChangeKind to
  allow automatic vote copy in this case
Change-Id: I23dfdfc42aa2f08fd647fc0dd1b5e2e45ea5dad8