EventFactory: Re-use revwalk for loading ChangeKindCache
We may have already parsed the commit with the already opened RevWalk or
we will later parse the commit with it. Either way, we can avoid
duplication of effort by passing it through to this cache.

This also skips attempting to re-open the repository and uses the
repoConfig every caller already has access to.

Release-Notes: skip
Change-Id: Ib6a5a6a44d64dab7b6da7d96c9b2a9b4500a26a2