Fix missing Submit ACL on dependant changes
The MergeOp logic doesn't check Submit permissions for dependant
changes. This can cause an issue with topics, since changes in the same
topic are not necessarily in the same branch (or even project).

This issue was caused by inconsistancy in checks between Submit UiAction
and MergeOp. To avoid similar issues in the future, instead of simply
fixing the check in MergeOp, we update both Submit.java and MergeOp to
use a shared implementation.

As the result of this, READ permission is potentially checked multiple
times. However there are potential reasons why it can be useful
(staleness) and trying to avoid it would add more complexity to the
change.

The check in Submit.java used to have special logic that uses current
ChangeData, instead of potentially stale information in ChangeSet
(because it's generated from index). To maintain the behaviour we
replace the data for the triggering change in the ChangeSet before
calling checkCommonSubmitProblems.

countChangesThatWereSubmittedWithRebaserApproval was moved out of
checking logic and now only counts the changes, if all changes passed
the checks. Which seems more inline with the intention based on the
name.

The Submit permission error for dependent changes is surfaced as
ResourceConflictException (as opposed to AuthException), since the error
message can contain multiple problems from commitStatus, and it's not
worth having a custom logic to "sometimes" have better error code.

Google-Bug-Id: b/347935402
Release-Notes: skip
Change-Id: Ic6dbda6dec1deebb6c14bb9056e0c20db768a9e1