web-dev-server: fix cors not being set
When using a development cdn such as:

  java -jar gerrit.war daemon --console-log \
    --site-path . \
    --dev-cdn http://localhost:8081/

The browser will complains about unallowed cross origin requests being
made. Some of the requests originates from `<script>` elements for which
the browser does not set the `Origin` header. In this case, the
@koa/cors does not set any CORS header.

The fix is https://github.com/koajs/cors/pull/87 which was released in
3.4.2, rolled back in 3.4.3 and rereleased as a major version 4.0.0.

5.0.0 includes a security fix to no more default the CORS header to be
automatically set to the requested origin since that defeats its
purpose. https://github.com/advisories/GHSA-qxrj-hx23-xp82

Release-Notes: skip
Change-Id: Ia2850e7fd88b4a20acfa6344469db69dafe6cb25