web-dev-server: serve some missing static assets
webcomponents-lite.js is exposed by index.html under /bower_components/
which is under the root dir when developping. When building Gerrit, the
file is copied by polygerrit_bundle().
Redirect to the in-tree files installed by yarn.

The fonts are also outside of the root dir, stream them from the
filesystem.

Release-Notes: skip
Change-Id: I10ad8a27bab2e5392b4c276346451caa6ec008c5