Fix Attach AI fixes on Save Button click
Change 420297 introduced bug what AI fixes are not added to draft
object when user clicks on Save button; they are added only by
auto-save, so it works when auto-save is triggered, but not on quick
comments where user clicks on save. This change removes
the buggy code that caused the issue of not attaching AI fixes
after user clicks on Save.

Save() is setting editing to false before calling somethingToSave()
which calls isFixSuggestionChanged(). So editing is always false
in isFixSuggestionChanged called by save() and it makes no sense to
return false when !this.editing in isFixSuggestionChanged().

Added unit tests that check adding AI fixes on user clicks. One fails
without this change.

Google-Bug-Id: b/343632982
Release-Notes: skip
Change-Id: Ic404fa17285d97b56b2f8d222f064b9f0780ec67