Make new image diff less strict
As noted in previous code, Resemble.js defaults to a loose difference by
setting some tolerance values for colors, alpha and min/max brightness.
Pending the introduction of a control to switch between the modes, the
default was to ignore `nothing` which highlights any changes.

Change the default to a loose diff (ignoreLess()), matching upstream
library default.

When 'Highlight differences' is checked, appears a set of radio buttons
mapped to ResembleJS built-in ignore options. Pseudo UI:

  [✓] Highlight differences
    ( ) All differences
    (*) Only substantial differences
    ( ) Ignore colors changes
    ( ) Ignore alpha channel changes
    ( ) Ignore anti-aliasing changes

I found the loose diff to be the most convenient in a lot of case and
one can then easily switch to `Nothing` to get the previous behavior.

Screenshots: https://phabricator.wikimedia.org/T341291#9939660

This change is for the UiFeature__new_image_diff_ui experiment.

The commit was previously reverted because Google internally are still
using Resemble.js 2 and the change could not be imported.

Release-Notes: Allow picking different ignore mode in UiFeature__new_image_diff_ui experiment view
Change-Id: I32a49d9fcc5cf9c5646280cf17a7709b616bd2c0