Make new image diff less strict
As noted in previous code, Resemble.js defaults to a loose difference by
setting some tolerance values for colors, alpha and min/max brightness.
Pending the introduction of a control to switch between the modes, the
default was to ignore `nothing` which highlights any changes.

Change the default to a loose diff (ignoreLess()), matching upstream
library default.

When 'Highlight differences' is checked, appears a set of radio buttons
mapped to ResembleJS built-in ignore options. Pseudo UI:

  [✓] Highlight differences
  Ignore:
    ( ) Nothing
    (*) Less
    ( ) Anti-aliasing
    ( ) Colors
    ( ) Alpha

I found the loose diff to be the most convenient in a lot of case and
one can then easily switch to `Nothing` to get the previous behavior.

Screenshots: https://phabricator.wikimedia.org/T341291#9939660

This change is for the UiFeature__new_image_diff_ui experiment.

Change-Id: Ie09b2c2b85991e875b7753430996f9d97fd1b296