Change new image diff API to ease configuration
The Resemble.js Comparison generated by `compareTo()` can be further
configured by chaining methods with preset settings such as
`ignoreNothing()`. It is a bit inconvenient to later implement a way fo
the user to pick which preset they want to apply.

There is an other interface, `compare()`, which can be passed a set of
options as string which makes it easier to bind a UI element to a
setting.

Change the Resemble.js call from compareTo to compare.  The chained
ignoreNothing() call is replaced with the equivalent comparison option
{ignore: 'nothing'}.

This change is for the UiFeature__new_image_diff_ui experiment.

Release-Notes: skip
Change-Id: Iee772b7cb2ebcaedd29dd2a91bd71f5ab3b53479