Resolve competition of comment range color and token color
The typical user interaction for range commenting is to drag the mouse
to select and then press `c`, and then start typing without further
mouse movement. This results in the token highlighting still being
applied while the user types the comment. If the user happens to only
comment on a substring of a token, then they cannot see the comment
range at all while typing the comment. This was reported in the
referenced issue.

Prefering the background color of the comment range highlight is a bit
tricky, because the highlights are not applied to the same element, and
the token element is always the inner element, and thus "wins". So we
have to adapt the css rule to make sure that there is not comment
highlighting element between `gr-diff-text` and the token.

We are also adapting the comment range hightlighting colors a tiny bit,
because it was quite impossible to see the difference of the token
highlight color and `--diff-highlight-range-hover-color` in light theme.

Release-Notes: skip
Google-Bug-Id: b/349651135
Change-Id: I26731a60a9e24af0d101ec5072b5188c1366a095