ReceiveCommits: Increment reject count metric for exceptions
Errors that are happening in the Git layer when updating refs are
causing exceptions (e.g. GitUpdateFailureException for us at Google).
insertChangesAndPatchSets catches the exceptions and wraps them first in
a RestApiException, then in a StorageException. StorageException is a
RuntimeException, hence to increase the reject count metric we catch
RuntimeExceptions in processCommands and then we rethrow the exception.
To format the cause we use ExceptionHook, the same as we do in
RetryHelper and RestApiServlet.

Bug: Google b/151127672
Release-Notes: skip
Change-Id: Iaba330ae9614198a879371b1f6fc378424a6b310
Signed-off-by: Edwin Kempin <ekempin@google.com>