ReceiveCommits: Log RuntimeException on severe level
Having this logged always is useful to investigate exceptions that are
counted by the reject count metric. If you see errors for an exception
in the reject count metric, and there is no formatting for this
exception yet, you would like to lookup and inspect the stacktrace and
this log allows this.

Bug: Google b/348209481
Change-Id: I14ae3f0567444335336749c13534d5373191bcd0
Signed-off-by: Edwin Kempin <ekempin@google.com>
Release-Notes: skip