ReceiveCommits: Call ExceptionHook to get status for exceptions
Not all exceptions that we catch in ReceiveCommits#processCommands
represent server errors. Call the ExceptionHooks to get the status for
the exception so that we can record the correct status in the reject
count metric.

If an exception doesn't represent a server error, log it only on INFO
severity.

Release-Notes: skip
Bug: Google b/151127672
Change-Id: Ib51147ea17e4340297541e59f0decfb424febf5b
Signed-off-by: Edwin Kempin <ekempin@google.com>