Specify order in the related changes tooltip
The order of parent/child in related changes is not always clear, there
are other improvements that can be done to improve the interface, but in
the meantime we can place this information in the tooltip. Especially
since currently the tooltip unhelpfully repeats name of the category.

Screenshot: https://imgur.com/a/xiyKc0F
Google-Bug-Id: b/352584416
Release-Notes: skip
Change-Id: I0a33e4bed74cc4e93c5ad26b94c50fce065df4b3