Add gr-focus-layer to gr-diff
The gr-focus-layer is added to the gr-diff component. This layer is
responsible for applying the  class to  elements within  components
when the corresponding line is out of focus.

The layer is partially based on the gr-coverage layer. For each line,
the layer checks whether it is in focus or out of focus. If it is out of
focus, the diff colors are saturated, and no green or red colors are
rendered for those diffs.

If is undefined, the layer is treated as disabled, and no logic is applied.

Change-Id: I4074e8e3eab90284243c284cb1c9b626b4628c42
Screencast: https://screencast.googleplex.com/cast/NDY2NTU4MDEwMjYxNTA0MHxjOGM4ZjBhOC0yMA
Release-Notes: skip
Google-Bug-Id: b/345126277