Disallow tracing configs that trigger tracing for too many requests
Misconfigured tracing configs or tracing configs that are too broad can
trigger tracing for too many requests. This can badly impact the
performance and the availability of the server, if tracing makes
requests more expensive (e.g. if there are expensive logs that are only
written when a request is traced).

Examples for a misconfigured or trace configs that are too broad:
* a tracing config that specifies neither of the match criteria (e.g. it
  contains an unsupported parameter, which can happen is the parameter
  has typo)
* a tracing config that only specifies match criteria that are very
  broad (e.g. requestType=REST which would trigger tracing for all REST
  requests)

Fix this by requiring that tracing configs specify at least one of
requestUriPattern, account and projectPattern.

Release-Notes: Disallowed tracing configs that trigger tracing for too many requests
Bug: Google b/355393231
Change-Id: I9c9fac92f45e427087f7fe705169f73296c0b2c8
Signed-off-by: Edwin Kempin <ekempin@google.com>