Fix onBehalfOf behaviour of Submit
When submitting changes using onBehalfOf only permission of the user
triggering the Submit need to be considered. With an Ic6dbda6de that
logic got broken, it was no longer possible to submit changes on behalf
of another user, if that user didn't have the SUBMIT permission
themselves.

This change fixes this issue by passing both users into MergeOp and
using either activeUser or mergeAsUser appropriately based on the
situation.

Additionally as part of this change we address the pre-existing issue,
where the SUBMIT_AS permission for active user and READ permission for
on-behalf-of user would only be checked for the triggering change and
not for all changes in the Submission. This lead to potential cases of
submitting changes as part of the topic that they would otherwise lack
permissions to submit. The extra checks for the "current change" in
Submit.java are kept, to allow the possibility of an early exit, as
construction of the full Set of changes to be submitted togetehr is an
expensive task.

By passing both activeUser and mergeAsUser to MergeOp we are able to
address another issue, where only READ permission of the onBehalfOf user
was considered when constructing Merge Set. With this change we use
activeUser's permission for constructing the Merge Set, but also
validate mergeAsUsers's READ permissions as well once it's constructed.

We also fix the permission check in ProjectConfigValidator, where
onBehalfOf user would be previously checked instead of the real user
performing the Submit.

It's unknown if plugin provided MergeValidationListener instances
require active or mergeAs user information, so we update the signature
to provide both. (Default implementation was added to avoid breaking
existing plugins)

Release-Notes: skip
Google-Bug-Id: b/351138952
Change-Id: Ib126d5fb75ab46620e302d35aa4a75699739732c