Drop remaining debug logs for known groups
Getting known groups is expensive. At the moment we log them when
requests are traced which makes traced requests more expensive. This
badly affects the performance if traced requests log known groups for
many users (e.g. the log in ExternalUser and the logs in IdentifiedUser
that we already dropped in change Ia2e7d8395). If now tracing is enabled
for too many requests this can bring down the server. To prevent this
from happening, we drop these logs. The logging of known groups in
RestApiServlet, ReceiveCommits and DefaultRefFilter is less bad as it
happens only once per requests (and likely known groups are needed at a
later point anyway), but to be on the save side we drop them as well.

Release-Notes: skip
Bug: Google b/354875955
Bug: Google b/355393768
Change-Id: I15647018c935adf51fe4fedd16e5831615826a38
Signed-off-by: Edwin Kempin <ekempin@google.com>